Load in angular files through the bundle
With any files that only interface through angular, we can simply switch them over to loading through webpack instead of the index page. An important thing to note is that this isn't going to work for files with globals. Once the file is imported this way, it becomes an es6 module file, and declarations are no longer considered globals. For this reason, we can't simply move utility.js over to this strategy.