Once all of our angular-dependent files are in the index bundle, we can move angular to the vendor bundle
Unlike our utility.js, this still works as a global since angular explicitly sets itself on the window. (We could do this too, but that doesn't bring us closer to our end goal of modernizing everything)

Note that we couldn't move angular to the vendor bundle until everything that depends on it was in the main bundle since webpack injects the two bundles at the bottom of the index page.