Setup a test config to run tests with webpack
Now that we're running the app through a webpack loader, we also need a webpack test loader, since our files are now es modules. Turns out we don't actually need much configuration. Just need to require in the files we need through the test setup file. We can use a wildcard to require in all spec files, like we did in the karma config before.