Convert the time service to angular 2
This is just a simple to 'downgrade' as the component was. When converting a service that uses http, you'll most likely need to switch to the new http service as well (unless you upgrade $http to make it available to angular 2). The new http module uses observables, which will require a change to all of its consumers. Alternatively, you can .toPromise() to turn the stream into an es6 promise for backwards compatibility. The types won't match the $http promise types, but consuming code should still be able to use it the same.