Drop in angular 2 routing alongside angular 1
In order to make this work, we'll need to drop the hash prefix in angular 1. This is because the routes need to match up in format between angular 1 and 2, and I'm not sure how to add the exclamation in angular 2. (You would probably need a location strategy for it.

Also, since we have an 'otherwise' route setup in angular 1, we need to define empty routes matching all the routes we have in angular 2. And finally, we need an app base href.