Converted the home route to angular 2
This is more interesting, since we now want to trigger an initial navigation on the new router, to allow for a redirect if '/' is hit. We can now drop the 'otherwise' in angular 1 and instead implement a home route in angular 2 with a redirect to it. Note that broken links will now result in errors instead of going to the home route. This is probably preferrable anyway.